feat: 가이드 이미지 추가 - 코드리뷰 반영 - 공지사항 빈 경우 가이드 이미지 추가
