+++ .eslintrc
@@ -31,6 +31,7 @@
       2,
       { "extensions": [".js", ".jsx", ".ts", ".tsx"] }
     ],
+    "react/require-default-props": 0,
     "@typescript-eslint/explicit-module-boundary-types": 0,
     "array-bracket-spacing": ["error", "always", { "arraysInArrays": true }],
     "lines-between-class-members": [
