+++ packages/mobile/src/__mocks__/index.ts
@@ -2,7 +2,17 @@ import constructionList from "./constructionList";
 import foodList from "./foodList";
 import imageList from "./imageList";
 import menuList from "./menuList";
+import { mockNotification, mockCategory } from "./notificationList";
 import playList from "./playList";
 import snackList from "./snackList";
 
-export { constructionList, foodList, imageList, menuList, playList, snackList };
+export {
+  constructionList,
+  foodList,
+  imageList,
+  menuList,
+  playList,
+  snackList,
+  mockNotification,
+  mockCategory,
+};
