+++ packages/mobile/src/__mocks__/notificationList.ts
@@ -1,0 +1,36 @@
+const mockCategory = [
+  { major: "경영정보학과", to: "?major=경영정보학과" },
+  { major: "소프트웨어학과", to: "?major=소프트웨어학과" },
+  { major: "안전공학과", to: "?major=안전공학과" },
+  { major: "어떤학과", to: "?major=어떤학과" },
+  { major: "이런학과저런학과", to: "?major=이런학과저런학과" },
+];
+
+const mockNotification = [
+  {
+    id: 1,
+    category: "충북대학교",
+    title: "충북대학교 인문대학 영상 공모전 안내1",
+    date: "22-04-22",
+    view: 123,
+    star: 55,
+  },
+  {
+    id: 2,
+    category: "충북대학교",
+    title: "충북대학교 인문대학 영상 공모전 안내2",
+    date: "22-04-22",
+    view: 123,
+    star: 55,
+  },
+  {
+    id: 3,
+    category: "충북대학교",
+    title: "충북대학교 인문대학 영상 공모전 안내3",
+    date: "22-04-22",
+    view: 123,
+    star: 55,
+  },
+];
+
+export { mockCategory, mockNotification };
