+++ packages/mobile/src/components/atoms/icon/Info.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/require-default-props */
 import type { StyleProps } from "src/type/props";
 
 type Props = {
@@ -9,6 +8,8 @@ type Props = {
 function Info({ className, style, width = "10", height = "10" }: Props) {
   return (
     <svg
+      className={className}
+      style={style}
       width={width}
       height={height}
       viewBox="0 0 10 10"
