+++ packages/mobile/src/components/atoms/icon/Setting.tsx
@@ -1,11 +1,17 @@
 import type { StyleProps } from "src/type/props";
 
-function Setting({ className, style }: StyleProps) {
+type Props = {
+  width?: string;
+  height?: string;
+} & StyleProps;
+
+function Setting({ className, style, width, height }: Props) {
   return (
     <svg
+      className={className}
       style={style}
-      width="24"
-      height="26"
+      width={width || 20}
+      height={height || 20}
       viewBox="0 0 24 26"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
